requirements.txt
asgiref==3.3.4
click==8.0.1
fastapi==0.65.2
h11==0.12.0
httptools==0.2.0
jinja2==3.0.1
markupsafe==2.0.1
pydantic==1.8.2
python-codicefiscale==0.3.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-slugify==5.0.2
pyyaml==5.4.1
six==1.16.0
starlette==0.14.2
text-unidecode==1.3
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
